requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
distlib==0.3.6
Django==4.1.9
django-cors-headers==3.14.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
filelock==3.12.0
idna==3.4
platformdirs==3.5.0
pycodestyle==2.10.0
pymongo==3.11.0
pytz==2023.3
requests==2.30.0
sqlparse==0.2.4
tzdata==2023.3
urllib3==2.0.2
virtualenv==20.23.0
